Fix accounting and permissions bugs

+++ database/migrations/2026_06_27_100005_create_productos_table.php
@@ -19,9 +19,9 @@ return new class extends Migration
             $table->foreignId('categoria_id')->constrained('categorias')->restrictOnDelete();
             $table->foreignId('unidad_medida_id')->constrained('unidades_medida')->restrictOnDelete();
             $table->decimal('precio_unitario', 12, 2)->default(0);
-            $table->integer('stock_minimo')->default(0);
-            $table->integer('stock_maximo')->default(0);
-            $table->integer('stock_actual')->default(0);
+            $table->decimal('stock_minimo', 12, 2)->default(0);
+            $table->decimal('stock_maximo', 12, 2)->default(0);
+            $table->decimal('stock_actual', 12, 2)->default(0);
             $table->string('estado', 20)->default('activo');
             $table->text('notas')->nullable();
             $table->timestamps();
